Fixes a small bug where I was using unsigned when transmitting movement packets so you couldn't see other people. Make it so server loads the map from file, sends the map to the users. Updates all users when a tile is changed by another user, and saves maps on server closeout with CTRL+C. This closes issue #5.

+++ Common/Src/Players.hpp
@@ -38,7 +38,7 @@ class cPlayer
 			_Coord.Y -= MOVE_AMOUNT * cos(Angle);
 			_Coord.X += MOVE_AMOUNT * sin(Angle);
 		}
-		void SetCoord(u32 X, u32 Y)
+		void SetCoord(f32 X, f32 Y)
 		{
 			_Coord.Y = Y;
 			_Coord.X = X;
